create component folder/file
