+++ client/front_end/src/component/collapsible_list.js
@@ -1,0 +1,26 @@
+import useCollapse from "react-collapsed";
+
+function Section(props) {
+  const config = {
+    defaultExpanded: props.defaultExpanded || false,
+    collapsedHeight: props.collapsedHeight || 0,
+  };
+  const { getCollapseProps, getToggleProps, isExpanded } = useCollapse(config);
+  return (
+    <div className="collapsible">
+      <div className="header" {...getToggleProps()}>
+        <div className="title">{props.title}</div>
+        <div className="icon">
+          <i
+            className={"fas fa-chevron-circle-" + (isExpanded ? "up" : "down")}
+          ></i>
+        </div>
+      </div>
+      <div {...getCollapseProps()}>
+        <div className="content">{props.children}</div>
+      </div>
+    </div>
+  );
+}
+
+export default Section;
